feat: SJRA-1339 adjust field types

+++ backend/internal/types/germline_snv_occurrence.go
@@ -190,7 +190,7 @@ var GermlineSNVAdRatioField = Field{
 	CanBeSelected: true,
 	CanBeFiltered: true,
 	CanBeSorted:   true,
-	Type:          IntegerType,
+	Type:          DecimalType,
 	Table:         GermlineSNVOccurrenceTable,
 }
 var GermlineSNVAdAltField = Field{
